+++ src/Common/AssemblyInfo.cs
@@ -4,17 +4,17 @@ using System.Reflection;
 
 [assembly: AssemblyProduct("FSharp.Formatting")]
 [assembly: AssemblyDescription("A package of libraries for building great F# documentation, samples and blogs")]
-[assembly: AssemblyVersion("6.0.1")]
-[assembly: AssemblyFileVersion("6.0.1")]
-[assembly: AssemblyInformationalVersion("6.0.1")]
+[assembly: AssemblyVersion("6.0.2")]
+[assembly: AssemblyFileVersion("6.0.2")]
+[assembly: AssemblyInformationalVersion("6.0.2")]
 [assembly: AssemblyCopyright("Apache 2.0 License")]
 namespace System {
     internal static class AssemblyVersionInformation {
         internal const System.String AssemblyProduct = "FSharp.Formatting";
         internal const System.String AssemblyDescription = "A package of libraries for building great F# documentation, samples and blogs";
-        internal const System.String AssemblyVersion = "6.0.1";
-        internal const System.String AssemblyFileVersion = "6.0.1";
-        internal const System.String AssemblyInformationalVersion = "6.0.1";
+        internal const System.String AssemblyVersion = "6.0.2";
+        internal const System.String AssemblyFileVersion = "6.0.2";
+        internal const System.String AssemblyInformationalVersion = "6.0.2";
         internal const System.String AssemblyCopyright = "Apache 2.0 License";
     }
 }