+++ src/FSharp.Formatting.Literate/Evaluator.fs
@@ -12,8 +12,15 @@ open FSharp.Formatting.Internal
 /// Represents a kind of thing that can be embedded 
 [<RequireQualifiedAccessAttribute>]
 type FsiEmbedKind = 
-  | Output
+  /// The FSI output 
+  | FsiOutput
+  /// The combined FSI output and console output
+  | FsiMergedOutput
+  /// The stdout from this part of the execution (not including FSI output)
+  | ConsoleOutput
+  /// The 'it' value
   | ItValue
+  /// A specific value
   | Value
 
 /// An interface that represents FSI evaluation result
@@ -24,6 +31,8 @@ type IFsiEvaluationResult = interface end
 /// the generated console output together with a result and its static type.
 type FsiEvaluationResult = 
   { Output : string option
+    FsiOutput : string option
+    FsiMergedOutput : string option
     ItValue : (obj * Type) option
     Result : (obj * Type) option }
   interface IFsiEvaluationResult
@@ -99,19 +108,242 @@ type FsiEvaluatorConfig() =
   static member CreateNoOpFsiObject() = box (new NoOpFsiObject())
 
 /// A wrapper for F# interactive service that is used to evaluate inline snippets
-type FsiEvaluator(?options:string[], ?fsiObj) =
-  // Initialize F# Interactive evaluation session
-  let fsiOptions = (Option.map FsiOptions.ofArgs options) |> Option.defaultWith (fun _ -> FsiOptions.Default)
-  let fsiSession = ScriptHost.Create(fsiOptions, preventStdOut = true, ?fsiObj = fsiObj)
+type FsiEvaluator(?options:string[], ?fsiObj: obj, ?addHtmlPrinter: bool, ?discardStdOut: bool, ?disableFsiObj: bool) =
+
+  let discardStdOut = defaultArg discardStdOut true
+  let fsiObj = defaultArg fsiObj (box FSharp.Compiler.Interactive.Shell.Settings.fsi)
+  let addHtmlPrinter = defaultArg addHtmlPrinter true
+  let disableFsiObj = defaultArg disableFsiObj false
+
+  let fsiOptions = options |> Option.map FsiOptions.ofArgs |> Option.defaultWith (fun _ -> FsiOptions.Default)
+  let fsiOptions =
+     if addHtmlPrinter then
+         { fsiOptions with Defines=fsiOptions.Defines @ [ "HAS_FSI_ADDHTMLPRINTER" ] }
+     else
+         fsiOptions
+  let fsiSession = ScriptHost.Create(fsiOptions, discardStdOut = discardStdOut, fsiObj = fsiObj)
+
+  let mutable plainTextPrinters : Choice<(obj -> string option), (obj -> obj option)>  list = []
+  let mutable htmlPrinters : Choice<(obj -> (seq<string * string> * string) option), (obj -> obj option)>  list = []
+
+  //----------------------------------------------------
+  // Inject the standard 'fsi' script control model into the evaluation session
+  // without referencing FSharp.Compiler.Interactive.Settings (which is highly problematic)
+  //
+  // Injecting arbitrary .NET values into F# interactive sessions from the outside is non-trivial.
+  // The technique here is to inject a script which reads values out of static fields
+  // in this assembly via reflection.
+
+  let thisTypeName = typeof<FsiEvaluator>.AssemblyQualifiedName
+  let addPrinterThunk(f: obj, ty: Type) =
+      let realPrinter (value:obj) =
+        match value with
+        | null -> None
+        | _ ->
+            if ty.IsAssignableFrom(value.GetType()) then
+                match f with
+                | :? (obj -> string)  as f2 ->
+                    match f2 value with
+                    | null -> None
+                    | s -> Some s
+                | _ -> None
+            else None
+      plainTextPrinters <- Choice1Of2 realPrinter :: plainTextPrinters
+
+  let addPrintTransformerThunk(f: obj, ty: Type) =
+      let realPrinter (value:obj) =
+        match value with
+        | null -> None
+        | _ ->
+            if ty.IsAssignableFrom(value.GetType()) then
+                match f with
+                | :? (obj -> obj)  as f2 ->
+                    match f2 value with
+                    | null -> None
+                    | o -> Some o
+                | _ -> None
+            else None
+      plainTextPrinters <- Choice2Of2 realPrinter :: plainTextPrinters
+      htmlPrinters <- Choice2Of2 realPrinter :: htmlPrinters
+
+  let addHtmlPrinterThunk(f: obj, ty: Type) =
+      let realHtmlPrinter (value:obj) =
+        match value with
+        | null -> None
+        | _ ->
+            if ty.IsAssignableFrom(value.GetType()) then
+                match f with
+                | :? (obj -> seq<string * string> * string) as f2 -> Some(f2 value)
+                | _ -> None
+            else None
+      htmlPrinters <- Choice1Of2 realHtmlPrinter :: htmlPrinters
+
+  let fsiEstablishText =
+      sprintf """
+namespace global
+[<AutoOpen>]
+module __FsiSettings =
+   open System
+   open System.Reflection
+   type fsi private() =
+       static let ty = System.Type.GetType("%s")
+       static let __InjectedAddPrinter =
+           ty.InvokeMember("InjectedAddPrinter", (BindingFlags.Static ||| BindingFlags.Public ||| BindingFlags.NonPublic ||| BindingFlags.GetProperty), null, null, [| |]) 
+           :?> ((obj -> string) * Type -> unit)
+       static let __InjectedAddPrintTransformer =
+           ty.InvokeMember("InjectedAddPrintTransformer", (BindingFlags.Static ||| BindingFlags.Public ||| BindingFlags.NonPublic ||| BindingFlags.GetProperty), null, null, [| |]) 
+           :?> ((obj -> obj) * Type -> unit)
+       static let __InjectedAddHtmlPrinter =
+           ty.InvokeMember("InjectedAddHtmlPrinter", (BindingFlags.Static ||| BindingFlags.Public ||| BindingFlags.NonPublic ||| BindingFlags.GetProperty), null, null, [| |]) 
+           :?> ((obj -> seq<string * string> * string) * Type -> unit)
+       static let __FsiObj =
+          ty.InvokeMember("InjectedFsiObj", (BindingFlags.Static ||| BindingFlags.GetProperty ||| BindingFlags.Public ||| BindingFlags.NonPublic), null, null, [| |]) 
+
+       static let getInstanceProp nm = __FsiObj.GetType().InvokeMember(nm, (BindingFlags.Instance ||| BindingFlags.GetProperty ||| BindingFlags.Public), null, __FsiObj, [|  |])
+       static let setInstanceProp nm v = __FsiObj.GetType().InvokeMember(nm, (BindingFlags.Instance ||| BindingFlags.SetProperty ||| BindingFlags.Public), null, __FsiObj, [| box v |]) |> ignore
+
+       /// Get or set the floating point format used in the output of the interactive session.
+       static member FloatingPointFormat
+           with get() = getInstanceProp "FloatingPointFormat" :?> string
+           and set(v:string) = setInstanceProp "FloatingPointFormat" v
+
+       /// Get or set the format provider used in the output of the interactive session.
+       static member FormatProvider 
+           with get() = getInstanceProp "FormatProvider" :?> System.IFormatProvider 
+           and set(v:System.IFormatProvider) = setInstanceProp "FormatProvider" v
+
+       /// Get or set the print width of the interactive session.
+       static member PrintWidth 
+           with get() = getInstanceProp "PrintWidth" :?> int  
+           and set(v:int  ) = setInstanceProp "PrintWidth" v
+
+       /// Get or set the print depth of the interactive session.
+       static member PrintDepth 
+           with get() = getInstanceProp "PrintDepth" :?> int
+           and set(v:int) = setInstanceProp "PrintDepth" v
+
+       /// Get or set the total print length of the interactive session.
+       static member PrintLength 
+           with get() = getInstanceProp "PrintLength" :?> int  
+           and set(v:int) = setInstanceProp "PrintLength" v
+
+       /// Get or set the total print size of the interactive session.
+       static member PrintSize  
+           with get() = getInstanceProp "PrintSize" :?> int
+           and set(v:int) = setInstanceProp "PrintSize" v
+
+       /// When set to 'false', disables the display of properties of evaluated objects in the output of the interactive session.
+       static member ShowProperties  
+           with get() = getInstanceProp "ShowProperties" :?> bool
+           and set(v:bool) = setInstanceProp "ShowProperties" v
+
+       /// When set to 'false', disables the display of sequences in the output of the interactive session.
+       static member ShowIEnumerable  
+           with get() = getInstanceProp "ShowIEnumerable" :?> bool
+           and set(v:bool) = setInstanceProp "ShowIEnumerable" v
+
+       /// When set to 'false', disables the display of declaration values in the output of the interactive session.
+       static member ShowDeclarationValues 
+           with get() = getInstanceProp "ShowDeclarationValues" :?> bool
+           and set(v:bool) = setInstanceProp "ShowDeclarationValues" v
+
+       /// Register a printer that controls the output of the interactive session.
+       static member AddPrinter<'T>(printer:'T -> string) =
+           __InjectedAddPrinter ((fun (v:obj) -> printer (unbox<'T> v)), typeof<'T>)
+           __FsiObj.GetType().GetMethod("AddPrinter", (BindingFlags.Instance ||| BindingFlags.Public)).MakeGenericMethod([| typeof<'T> |]).Invoke(__FsiObj, [|  box printer |]) |> ignore
+
+       /// Register a print transformer that controls the output of the interactive session.
+       static member AddPrintTransformer<'T>(printer: 'T -> obj) =
+           __InjectedAddPrintTransformer ((fun (v:obj) -> printer (unbox<'T> v)), typeof<'T>)
+           __FsiObj.GetType().GetMethod("AddPrintTransformer", (BindingFlags.Instance ||| BindingFlags.Public)).MakeGenericMethod([| typeof<'T> |]).Invoke(__FsiObj, [|  box printer |]) |> ignore
+
+       static member AddHtmlPrinter<'T>(printer:'T -> seq<string * string> * string) =
+           __InjectedAddHtmlPrinter ((fun (v:obj) -> printer (unbox<'T> v)), typeof<'T>)
+
+       /// The command line arguments after ignoring the arguments relevant to the interactive environment
+       static member CommandLineArgs 
+           with get() = getInstanceProp "CommandLineArgs" :?> string []
+           and set(v:string[]) = setInstanceProp "CommandLineArgs" v
+    """ thisTypeName
+
+  do
+    if not disableFsiObj then
+        FsiEvaluator.InjectedAddPrinter <- addPrinterThunk
+        FsiEvaluator.InjectedAddPrintTransformer <- addPrintTransformerThunk
+        FsiEvaluator.InjectedAddHtmlPrinter <- addHtmlPrinterThunk
+        FsiEvaluator.InjectedFsiObj <- fsiObj
+
+        let path = Path.GetTempFileName() + ".fs"
+        File.WriteAllText(path, fsiEstablishText)
+        let outputs, res = fsiSession.EvalInteraction(sprintf "#load @\"%s\"" path)
+        File.Delete(path)
+        match res with
+        | Ok v -> ()
+        | Error exn ->
+           printfn "Error establishing FSI:"
+           printfn "%s" outputs.Output.FsiOutput
+           printfn "%s" outputs.Output.ScriptOutput
+           printfn "%s" outputs.Error.FsiOutput
+           printfn "%s" outputs.Error.ScriptOutput
+           printfn "Exception: %A" exn
+           raise exn
+
 
   let evalFailed = new Event<_>()
   let lockObj = obj()
+
+  let rec plainTextPrint depth (v:obj) =
+      // guard against recursion in print transformers
+      if depth > 20 then
+          try sprintf "%A" v with e -> e.ToString() 
+      else
+          plainTextPrinters
+          // Try to find a printer or print transformer
+          |> List.tryPick (fun f ->
+               match f with
+               | Choice1Of2 addedPrinter -> 
+                  try addedPrinter v with _ -> None
+               | Choice2Of2 addedPrintTransformer ->
+                  (try addedPrintTransformer v with _ -> None)
+                  |> Option.map (plainTextPrint (depth+1)))
+          |> function
+              | None ->
+                  // no printer found
+                  try sprintf "%A" v with e -> e.ToString() 
+              | Some t -> t
+
+  let rec tryHtmlPrint depth (v:obj) =
+      if depth > 10 then
+          None
+      else
+          htmlPrinters
+          // Try to find a printer or print transformer
+          |> List.tryPick (fun f ->
+               match f with
+               | Choice1Of2 addedPrinter ->
+                  try addedPrinter v with _ -> None
+               | Choice2Of2 addedPrintTransformer ->
+                  (try addedPrintTransformer v with _ -> None)
+                  |> Option.bind (tryHtmlPrint (depth+1)))
 
   /// Registered transformations for pretty printing values
   /// (the default formats value as a string and emits single CodeBlock)
   let mutable valueTransformations = 
     [ (fun (o:obj, t:Type, executionCount: int) ->
-        Some([OutputBlock(sprintf "%A" o, "text/plain", Some executionCount)]) ) ]
+        tryHtmlPrint 0 o |> Option.map (fun (_tags, html) -> 
+           [OutputBlock(html, "text/html", Some executionCount)]) )
+
+      (fun (o:obj, t:Type, executionCount: int) ->
+        Some([OutputBlock(plainTextPrint 0 o, "text/plain", Some executionCount)]) ) ]
+
+  /// Temporarily holds the function value injected into the F# evaluation session
+  static member val internal InjectedAddPrintTransformer : ((obj -> obj) * Type -> unit) = Unchecked.defaultof<_>  with get, set
+  /// Temporarily holds the function value injected into the F# evaluation session
+  static member val internal InjectedAddPrinter : ((obj -> string) * Type -> unit) = Unchecked.defaultof<_>  with get, set
+  /// Temporarily holds the function value injected into the F# evaluation session
+  static member val internal InjectedAddHtmlPrinter : ((obj -> seq<string * string> * string) * Type -> unit) = Unchecked.defaultof<_>  with get, set
+  /// Temporarily holds the object value injected into the F# evaluation session
+  static member val internal InjectedFsiObj : obj = Unchecked.defaultof<_>  with get, set
 
   /// Register a function that formats (some) values that are produced by the evaluator.
   /// The specified function should return 'Some' when it knows how to format a value
@@ -128,8 +360,16 @@ type FsiEvaluator(?options:string[], ?fsiObj) =
       if not (result :? FsiEvaluationResult) then 
         invalidArg "result" "FsiEvaluator.Format: Expected 'FsiEvaluationResult' value as argument."
       match result :?> FsiEvaluationResult, kind with
-      | result, FsiEmbedKind.Output -> 
+      | result, FsiEmbedKind.ConsoleOutput -> 
         let outputText = defaultArg result.Output "No output has been produced."
+        let output = outputText.Trim()
+        [ OutputBlock (output, "text/plain", Some executionCount) ]
+      | result, FsiEmbedKind.FsiOutput -> 
+        let outputText = defaultArg result.FsiOutput "No output has been produced."
+        let output = outputText.Trim()
+        [ OutputBlock (output, "text/plain", Some executionCount) ]
+      | result, FsiEmbedKind.FsiMergedOutput -> 
+        let outputText = defaultArg result.FsiMergedOutput "No output has been produced."
         let output = outputText.Trim()
         [ OutputBlock (output, "text/plain", Some executionCount) ]
       | { ItValue = Some (obj, ty) }, FsiEmbedKind.ItValue
@@ -161,18 +401,33 @@ type FsiEvaluator(?options:string[], ?fsiObj) =
             | Some f -> Path.GetDirectoryName f
             | None -> Directory.GetCurrentDirectory()
           fsiSession.WithCurrentDirectory dir (fun () ->
-            let (output, value), itvalue =
+            let output, value, itvalue =
               if asExpression then
-                fsiSession.TryEvalExpressionWithOutput text, None
+                let output, res = fsiSession.TryEvalExpression text
+                match res with
+                | Error _ -> output, None, None
+                | Ok res -> output, res, None
               else
-                let output = fsiSession.EvalInteractionWithOutput text
-                // try get the "it" value, but silently ignore any errors
-                try
-                  (output, None), fsiSession.TryEvalExpression "it"
-                with _ -> (output, None), None
-            { Output = Some output.Output.ScriptOutput; Result = value; ItValue = itvalue  } :> _
+                let output, res = fsiSession.EvalInteraction text
+                match res with
+                | Error _ -> output, None, None
+                | Ok _ ->
+                    // try get the "it" value, but silently ignore any errors
+                    let _outputs, res = fsiSession.TryEvalExpression "it"
+                    match res with
+                    | Ok v -> output, None, v
+                    | Error _ -> output, None, None
+            { Output = Some output.Output.ScriptOutput
+              FsiMergedOutput = Some output.Output.Merged
+              FsiOutput = Some output.Output.FsiOutput
+              Result = value
+              ItValue = itvalue  } :> IFsiEvaluationResult
           )
       with :? FsiEvaluationException as e ->
         evalFailed.Trigger { File=file; AsExpression=asExpression; Text=text; Exception=e; StdErr = e.Result.Error.Merged }
-        { Output = None; Result = None; ItValue = None } :> _
+        { Output = None
+          FsiOutput = None
+          FsiMergedOutput = None
+          Result = None
+          ItValue = None } :> IFsiEvaluationResult