+++ RELEASE_NOTES.md
@@ -1,3 +1,12 @@
+## 6.0.2
+
+  - Always inject `fsi.AddPrinter`, `fsi.AddHtmlPrinter` etc. into the programming model for literate scripts
+  - Remove the `api` command from the command line tool (`build` generalises it)
+ - Add missing search.js
+ - Add `(*** include-fsi-output **)`
+ - Add `(*** include-fsi-merged-output **)`
+
+
 ## 6.0.1
 
   - build the Lunr `index.json` from every execution of `fsdocs build`